moved the map to components folder
